+++ t/100-refine-toml.t
@@ -1,6 +1,6 @@
 use v6.c;
 use Test;
-#use Data::Dump::Tree;
+use Data::Dump::Tree;
 use Config::DataLang::Refine;
 
 #-------------------------------------------------------------------------------
@@ -58,6 +58,8 @@ spurt( 'myCfg.cfg', Q:to/EOOPT/);
 
   [p3]
     name        = 'key=a string! &@data'
+    command     = '"touch "`date +%Y`.log'
+    c3tick      = '"`touch "`date +%Y`.log'
   EOOPT
 
 #-------------------------------------------------------------------------------
@@ -197,7 +199,7 @@ subtest {
 subtest {
 
   my Config::DataLang::Refine $c .= new(:config-name<myCfg.cfg>);
-#say dump $c;
+#say dump $c.config;
   my Array $o = $c.refine-str( <app>, :filter, :str-mode(C-UNIX-OPTS-T1));
   ok '--port=2345' ~~ any(@$o), 'app --port in list';
   nok '--workdir=/tmp' ~~ any(@$o), 'app --workdir /tmp not in list';
@@ -216,8 +218,12 @@ subtest {
   ok '--workdir=/var/tmp' ~~ any(@$o), 'app --workdir /var/tmp  in list';
 
   $o = $c.refine-str( <app p2>, :filter, :str-mode(C-UNIX-OPTS-T1));
-#say dump $o;
   ok "--text='abc def xyz'" ~~ any(@$o), 'p2 --text: spaced text';
+
+  $o = $c.refine-str( <p3>, :str-mode(C-UNIX-OPTS-T1));
+say dump $o;
+  ok '--command="touch "`date +%Y`.log' ~~ any(@$o), 'p3 --command: backtick text';
+  ok '--c3tick=\'"`touch "`date +%Y`.log\'' ~~ any(@$o), 'p3 --c3tick: backtick text';
 
   $o = $c.refine-str( <p2>, :filter, :str-mode(C-UNIX-OPTS-T1));
   ok '--workdir=~/p2' ~~ any(@$o), 'p2 --workdir ~/p2 in list';