+++ t/200-refine-json.t
@@ -238,7 +238,6 @@ subtest {
   ok '-pq' ~~ any(@$o), 'app -pq in list';
   ok '--notest' ~~ any(@$o), 'app -notest in list';
 
-
   $o = $c.refine-str( <app>, :filter, :str-mode(C-UNIX-OPTS-T2));
   ok '--port=2345' ~~ any(@$o), 'app --port in list';
   nok '--workdir=/tmp' ~~ any(@$o), 'app --workdir /tmp not in list';
@@ -246,7 +245,13 @@ subtest {
   nok '-q' ~~ any(@$o), 'app -q not in list';
   ok '-pq' ~~ any(@$o), 'app -pq in list';
 
-}, 'refine filter string array tests C-UNIX-OPTS-T2';
+  $o = $c.refine-str( <app p2>, :filter, :str-mode(C-UNIX-OPTS-T3));
+#say dump $o;
+  ok "--/test" ~~ any(@$o),
+     'p2 --/test: negated perl6 option, filter ignored';
+  ok "--/tunnel" ~~ any(@$o),
+     'p2 --/tunnel: negated perl6 option, filter ignored';
+}, 'refine filter string array tests C-UNIX-OPTS-T2 and T3';
 
 #-------------------------------------------------------------------------------
 # Cleanup